Improved expression eval and added int_bounds method to optimizers.
Also:
- reorganized bounds and relation types in to their own module,
- added % op to Int to aid better eval behaviour
- added div_ceil and div_floor to Int
- corrected documentation
