+++ src/proof/optimiser.rs
@@ -1,7 +1,7 @@
 use super::ast::Ident;
+use super::bound::Bound;
 use super::expr::{Atom, Expr};
 use super::int::Int;
-use super::Bound;
 use std::iter::Iterator;
 
 // TODO These types can probably be created with a macro to remove duplication.
@@ -11,8 +11,12 @@ use std::iter::Iterator;
 // efficient but make mistakes less likely. (Wouldn't attributes for this be so nice? We could
 // consider adding a phantom type to Expr and Bound also to garentee this.)
 
-/// Minimizer is a type for generating upper bounds on an expression, given a set of requirements.
+/// Minimizer is a type for generating lower bounds on an expression, given a set of requirements.
 /// It can be constructed with `Minimizer::new`, see that documentation for more details.
+///
+/// Note that upper bounds should be rounded up when evaluated.
+/// Rounding down will not lead to incorrect behaviour but can lead to loose bounds.
+/// This is the behaviur of the int_bounds() method.
 pub struct Minimizer<'a> {
     /// The expression to find bounds on
     /// This must be simplified.
@@ -30,8 +34,11 @@ pub struct Minimizer<'a> {
     child: Option<Box<Minimizer<'a>>>,
 }
 
-/// Maximizer is a type for generating lower bounds on an expression, given a set of requirements.
+/// Maximizer is a type for generating upper bounds on an expression, given a set of requirements.
 /// It can be constructed with `Maximizer::new`, see that documentation for more details.
+///
+/// Note that lower bounds should be down up when evaluated.
+/// This is the behaviur of the int_bounds() method.
 pub struct Maximizer<'a> {
     /// For field documentation, see Minimizer.
     /// Note that expressions and bounds must be simplified.
@@ -46,7 +53,7 @@ pub struct Maximizer<'a> {
 impl<'a> Minimizer<'a> {
     /// Returns a minimizer for `solve`.
     /// This minimizer will, using the known bounds given by `given`, iterate through expressions
-    /// such that `solve` <= `expr` for any valid values of named atoms.
+    /// such that `solve` >= `expr` for any valid values of named atoms.
     /// The resulting expressions will have varying degrees of tightness.
     /// The iterator ends when no more bounds are available.
     ///
@@ -73,12 +80,20 @@ impl<'a> Minimizer<'a> {
             child: None,
         }
     }
+
+    /// Returns an iterator of evaluated bounds.
+    /// So `solve >= x` for all x generated by this integer.
+    pub fn int_bounds(
+        self,
+    ) -> std::iter::FilterMap<Self, fn(Expr<'a>) -> std::option::Option<Int>> {
+        self.filter_map(|expr| expr.eval_ceil())
+    }
 }
 
 impl<'a> Maximizer<'a> {
     /// Returns a maximizer for `solve`.
     /// This maximizer will, using the known bounds given by `given`, iterate through expressions
-    /// such that `solve` >= `expr` for any valid values of named atoms.
+    /// such that `solve` <= `expr` for any valid values of named atoms.
     /// The resulting expressions will have varying degrees of tightness.
     /// The iterator ends when no more bounds are available.
     ///
@@ -104,6 +119,14 @@ impl<'a> Maximizer<'a> {
             max_depth,
             child: None,
         }
+    }
+
+    /// Returns an iterator of evaluated bounds.
+    /// So `solve <= x` for all x generated by this integer.
+    pub fn int_bounds(
+        self,
+    ) -> std::iter::FilterMap<Self, fn(Expr<'a>) -> std::option::Option<Int>> {
+        self.filter_map(|expr| expr.eval_floor())
     }
 }
